cleaned up code and removed unnecessary default params
