+++ app/src/main/java/com/cornellappdev/uplift/ui/components/goalsetting/ScrollWheel.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.gestures.snapping.rememberSnapFlingBehavior
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -20,7 +19,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -36,10 +34,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cornellappdev.uplift.util.GRAY01
-import com.cornellappdev.uplift.util.GRAY08
 import com.cornellappdev.uplift.util.montserratFamily
-import kotlinx.coroutines.launch
-import kotlin.math.abs
 
 /**
  * @param width: the width of the scroll wheel
@@ -132,81 +127,6 @@ fun <T> ScrollWheel(
         })
     }
 }
-
-/**
- * @param initialIsAm: whether the initial selection is AM
- * @param onValueChanged: callback for when the selection changes
- * @return AmPmSelector composable that allows the user to select AM or PM
- */
-@Composable
-fun AmPmSelector(
-    initialIsAm: Boolean = true, onValueChanged: (Boolean) -> Unit = {}
-) {
-    val items = listOf("AM", "PM")
-    val itemHeight = 24.dp
-    val scope = rememberCoroutineScope()
-    val scrollState = rememberLazyListState()
-    var selectedIndex by remember { mutableIntStateOf(if (initialIsAm) 0 else 1) }
-    val density = LocalDensity.current
-
-    LaunchedEffect(Unit) {
-        scrollState.scrollToItem(if (initialIsAm) 0 else 1)
-    }
-    Box(
-        modifier = Modifier.height(itemHeight * 3)
-    ) {
-        LazyColumn(
-            state = scrollState,
-            modifier = Modifier.width(24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            userScrollEnabled = true
-        ) {
-            // Top spacer
-            item {
-                Spacer(modifier = Modifier.height(itemHeight))
-            }
-
-            items(items.size) { index ->
-                val isSelected = selectedIndex == index
-                Box(
-                    modifier = Modifier
-                        .height(itemHeight)
-                        .fillMaxWidth()
-                        .onGloballyPositioned { coordinates ->
-                            val center = coordinates.size.height / 2f
-                            val position = coordinates.positionInParent().y + center
-                            val totalHeight = with(density) { (itemHeight * 3).toPx() }
-                            val middlePosition = totalHeight / 2f
-
-                            if (abs(position - middlePosition) <
-                                with(density) { itemHeight.toPx() } / 2
-                            ) {
-                                if (selectedIndex != index) {
-                                    selectedIndex = index
-                                    onValueChanged(index == 0)
-                                    scope.launch {
-                                        scrollState.animateScrollToItem(index)
-                                    }
-                                }
-                            }
-                        }, contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = items[index],
-                        fontSize = 14.sp,
-                        fontWeight = FontWeight.Bold,
-                        color = if (isSelected) Color.Black else GRAY08,
-                        fontFamily = montserratFamily
-                    )
-                }
-            }
-            item {
-                Spacer(modifier = Modifier.height(itemHeight))
-            }
-        }
-    }
-}
-
 
 @Preview(showBackground = true)
 @Composable
